+++ apps/web-customer/src/components/blocks/HubBlocks/ProgramListBlock/ProgramGridView.tsx
@@ -6,7 +6,7 @@ import {
   InProgressWhiteIcon,
   CompletedWhiteIcon,
 } from "core-library/assets";
-import { StandardProgramListType } from "../../../../core/types/programList";
+import { StandardProgramListType } from "core-library/types/wc/programList";
 import useCalculateProgramProgress from "../../../../core/hooks/useCalculateProgramProgress";
 import { ProgressCircle } from "../../../../components/ProgressCircle/ProgressCircle";
 import { DialogBox } from "core-library/components/Dialog/DialogBox";
