+++ apps/web-customer/src/components/blocks/HubBlocks/ProgramListBlock/ProgramListBlock.tsx
@@ -3,9 +3,15 @@ import { Box } from "@mui/material";
 import { AccordionList } from "../../../AccordionList/AccordionList";
 import { ProgramHeader } from "./ProgramHeader";
 import { ProgramGridView } from "./ProgramGridView";
-import { standardProgramList } from "../../../../core/constant/ProgramListMock/ProgramListMock";
+import { StandardProgramListType } from "core-library/types/wc/programList";
 
-export function ProgramListBlock(){
+interface ProgramListBlockProps {
+  program: StandardProgramListType[];
+  programTitle: string;
+  programSubtitle: string;
+}
+
+export const ProgramListBlock: React.FC<ProgramListBlockProps> = ({program, programTitle, programSubtitle}) => {
   const [listView, setListView] = useState<boolean>(true);
 
   const toggleView = () => setListView((prev) => !prev);  
@@ -14,18 +20,18 @@ export function ProgramListBlock(){
     <section className="flex h-auto w-full justify-center bg-[#e7eaf1]">
       <Box className="flex flex-col w-full lg:w-[1045px] mt-[120px] mb-[40px] mx-4 gap-8 ">
         <ProgramHeader
-          title={"Body Systems Topic (Day 1-13)"}
-          subtitle={"Key body systems covered over 13 days of study."}
+          title={programTitle}
+          subtitle={programSubtitle}
           listView={listView}
           handleClick={toggleView}
         />
         {listView ? (
           <div className="fadeIn">
-            <AccordionList program={standardProgramList} />
+            <AccordionList program={program} />
           </div>
         ) : (
           <>
-            <ProgramGridView program={standardProgramList} />
+            <ProgramGridView program={program} />
           </>
         )}
       </Box>
