+++ apps/web-customer/src/components/blocks/HubBlocks/ProgramListBlock/WatchVideosBlock/VideoPlayer.tsx
@@ -1,4 +1,4 @@
-import { SectionVideosType } from "../../../../../core/types/programList";
+import { SectionVideosType } from "core-library/types/wc/programList";
 import ReactPlayer from "react-player";
 import { useResolution } from "core-library/hooks";
 interface VideoPlayerProps {
