--- apps/web-customer/src/pages/hub/programs/index.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { ProgramListBlock } from "../../../components/blocks/HubBlocks/ProgramListBlock/ProgramListBlock";
-
-const ProgramListPage: React.FC = () => {
-  return <ProgramListBlock />;
-};
-
-export default ProgramListPage;
